requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.11.3
google-genai==1.12.1
